Initialization changes, additional CSR work
In addtion to existing initialization, initialize the XDM
command queue memory. After all initialization is done,
clear the ECC logs.

Changed the brige csr_mutex to probe_mutex and locked the
entire probe function. I haven't been able to determine
whether it is guaranteed the PCI layer prevents races or
not.

The driver will not open unless all slices are successfully
probed, an error occurred probe, or physical slice zero is
not present. The vslice mechanism is not working and
additional changes would be required to enable
driver-to-driver communication to function without physical
slice zero. The simplest fix would be to modify the UUID to
contain the slice of the driver's zero queue.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>